% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_overlap.R
\name{weight_overlap}
\alias{weight_overlap}
\title{Calculate the three-way dyadic population overlap between 2 levels of geography, with an atomic base layer}
\usage{
weight_overlap(
  shp1,
  shp_atom,
  shp2,
  crs1 = "+proj=laea +lat_0=10 +lon_0=-81 +ellps=WGS84 +units=m +no_defs",
  crs2 = "+init=epsg:2163",
  pop_field
)
}
\arguments{
\item{shp1}{The imported first shapefile, sp object}

\item{shp_atom}{The atomic unit of geography shpfile that will be used to calculate the spatial weights by population}

\item{shp2}{The imported second shapefile, sp object}

\item{crs1}{The CRS projection for the sf object transformation; default set to "+init=epsg:2163"}

\item{pop_field}{The field with the population data, from the shp_atom shp file.}
}
\value{
The data frame with the dyadic population overlap between the first and second shapefiles. The following are the values:
\itemize{
\item areal_weight = A 0-1 scaled weight of how much the intersection of the three shpfiles consists of a full atomic unit
\item pop_field = The renamed population field
\item pop_wt = The Calculated population within the dyad of shp1-atom_shp-shp2
}
}
\description{
This function takes the spatialDataframe objects part of the sp pkg and calculates the raw output of the three-way intersection
between two levels of geography of interest, overlaid ontop of an atomic shpfile layer, i.e. census blocks
}
\examples{

zctas <- arealOverlapr::zctas
cbg_oh <- arealOverlapr::cbg_oh
oh_sen <- arealOverlapr::oh_sen
test_overlap <-weight_overlap(shp1 = zctas, shp_atom = cbg_oh, shp2 = oh_sen, pop_field = "POP2010")
 )
}
